Build the scene dynamically using a Script
Create a new scene with only a camera (and, maybe, directional
lighting). Attach the Scripts/Initialize.cs script to the camera, and
the world is build dynamically. No more need to edit all those pesky
settings in Unity.

You control everything from the script file.

In addition, there is an AppStateSettings class that contains the center
coordinate (for now), which should make it easy to move the map using
voice commands.