Pois now visible & Placeable Anchor
POIS:
The UI layer has now been changed. The Pois are no longer added
underneath the UI but op the poi-Target rendered on the tile. This
solves the scaling and positioning  issues.
In additions POIS are no longer are GUI Image they are a 3d objects
which rotates with the camera angle.

Minor issue remains that if the building is to high the pois image is
inside...

Placeable Anchor:
The anchor is now placeable and a gaze cursor is added.
Consideration: Change the method of changing the anchor to a speech
command to prevent involuntary movement

Changes to scene: added prefab that needs to be added to the ini script.