Symbolfactory first version draggin with mouse only!
This is the first version of dragging the Symbols across the map.
It uses the cursor for the positioning making it impossible to drag the
symbols of the map.

To do further:
Make a command/ gesture to select
Make a button to start dragging
Make a command to stop dragging